Changes for upgrading to awesome 3.5
Notably the changes for the systray colors.

This is a hack to make sure that the icon colors in the systray are proper.

I do not pretend to understand why this is needed, but it works.

The numbers represent RGB values for the colors. The link in the code
leads to some mailinglist threads which explain the situation in a bit
more detail